desktop_launcher/requirements.txt
pywebview>=5.3,<6.0
